sketch-f/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["Alexander Mann-Wahrenberg <alexander.mannwahrenberg@gmail.com>"]
license = "CC BY-SA 4.0"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
grpcio = "^1.62.1"
python-dotenv = "^1.0.1"

[tool.poetry.group.dev.dependencies]
grpcio-tools = "^1.62.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
